Backend/servicios-IA-voz/voz-texto-api-ms/requirements.txt
fastapi>=0.100.0
uvicorn>=0.21.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.5
SpeechRecognition>=3.10.0
requests>=2.28.0
numpy>=1.24.0
pydub>=0.25.1
loguru>=0.7.0
pytest>=7.0.0
httpx>=0.24.0
imageio-ffmpeg>=0.4.8
jinja2>=3.1.2